% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_nje.R
\name{compute_nje}
\alias{compute_nje}
\title{Normalized jerk error}
\usage{
compute_nje(
  t,
  x,
  y,
  z,
  logging_frequency,
  enable_ksmooth = TRUE,
  included = rep(TRUE, length(t)),
  visualization = FALSE
)
}
\arguments{
\item{t}{A list of timestamps}

\item{x}{A list of x-axis positions/rotations during a movement}

\item{y}{A list of y-axis positions/rotations during a movement}

\item{z}{A list of z-axis positions/rotations during a movement}

\item{logging_frequency}{The logging frequency of your data}

\item{included}{A list of included timestamps, if you want to analyze a subset of the movement}

\item{visualization}{Whether to enable the built-in visualization}
}
\value{
The calculated NJE
}
\description{
Normalized jerk error
}
\examples{
library(MotorLearningMetrics4HCI)
df <- read.csv("exampledata.csv")
compute_nje(df$Timestamp, df$X, df$Y, df$Z, 50, visualization=TRUE)
}
