% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_npe.R
\name{compute_npe}
\alias{compute_npe}
\title{Normalized Path Error (NPE)}
\usage{
compute_npe(t, x, y, z, shortest_path_length, included = rep(TRUE, length(t)))
}
\arguments{
\item{t}{A list of timestamps}

\item{x}{A list of x-axis positions/rotations during a movement}

\item{y}{A list of y-axis positions/rotations during a movement}

\item{z}{A list of z-axis positions/rotations during a movement}

\item{shortest_path_length}{The pre-defined shortest path for the task condition}

\item{included}{A list of included timestamps, if you want to analyze a subset of the movement}
}
\value{
The calculated NPE
}
\description{
Normalized Path Error (NPE)
}
\examples{
library(MotorLearningMetrics4HCI)
df <- read.csv("exampledata.csv")
compute_npe(df$Timestamp, df$X, df$Y, df$Z, 0.24) #Note that there is no smoothing of positions in this implementation.
}
