% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_rtp.R
\name{compute_rtp}
\alias{compute_rtp}
\title{Refinement Time Proportion (RTP)}
\usage{
compute_rtp(
  t,
  x,
  y,
  z,
  enable_ksmooth = TRUE,
  peak_detection_algorithm = z_score_peak_detection,
  included = rep(TRUE, length(t)),
  visualization = FALSE
)
}
\arguments{
\item{t}{A list of timestamps}

\item{x}{A list of x-axis positions/rotations during a movement}

\item{y}{A list of y-axis positions/rotations during a movement}

\item{z}{A list of z-axis positions/rotations during a movement}

\item{enable_ksmooth}{Whether to enable the built-in kernel smoother}

\item{peak_detection_algorithm}{customized peak detection algorithm, if any}

\item{included}{A list of included timestamps, if you want to analyze a subset of the movement}

\item{visualization}{Whether to enable the built-in visualization}
}
\value{
The calculated RTP
}
\description{
Refinement Time Proportion (RTP)
}
\examples{
library(MotorLearningMetrics4HCI)
df <- read.csv("exampledata.csv")
compute_rtp(df$Timestamp, df$X, df$Y, df$Z, visualization = TRUE)
}
