% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_score_peak_detection.R
\name{z_score_peak_detection}
\alias{z_score_peak_detection}
\title{Z-score Peak Detection Algorithm}
\usage{
z_score_peak_detection(y, lag = 3, threshold = 3, influence = 0.1)
}
\arguments{
\item{y}{A list of timeseries data}

\item{lag}{The lag of the moving window that calculates the mean and standard deviation of historical data. Default = 3.}

\item{threshold}{The "z-score" at which the algorithm signals. Default = 3.}

\item{influence}{The influence (between 0 and 1) of new signals on the calculation of the moving mean and moving standard deviation. Default = 0.1.}
}
\value{
A list of values containing the signal results, the average filter, and the std filter. See https://stackoverflow.com/questions/22583391/peak-signal-detection-in-realtime-timeseries-data
}
\description{
Z-score Peak Detection Algorithm
}
\examples{
library(MotorLearningMetrics4HCI)
df <- read.csv("exampledata.csv")
compute_rtp(df$Timestamp, df$X, df$Y, df$Z, peak_detection_algorithm = z_score_peak_detection, visualization = TRUE)
}
